feat(painting): add complete chaining API for Canvas
Add fluent chaining methods for all Canvas drawing operations:

Drawing methods:
- texture() - GPU textures
- image(), image_repeat(), image_nine_slice(), image_filtered() - images
- shadow() - shadows
- gradient(), gradient_rrect() - gradients
- arc() - arcs
- drrect() - difference of rounded rects
- points(), vertices() - points and vertices
- rounded_rect() - rectangle with uniform corner radius

Updated RenderObjects to use modern Canvas API:
- paragraph.rs - draw_text() → text(), Paint::fill()
- texture.rs - draw_texture() → texture()
- error_box.rs - Paint::fill(), Paint::stroke(), rect(), line(), text()
- placeholder.rs - Paint::fill(), Paint::stroke(), rect(), text()
- physical_model.rs - shadow(), rounded_rect()
- physical_shape.rs - shadow()
- decorated_box.rs - rounded_rect()
- devtools.rs - Paint::fill(), Paint::stroke() (~18 places)

Documentation:
- Updated flui_painting/README.md with all new chaining methods
- Updated doc comments in render_box.rs, contexts.rs, custom_paint.rs

+++ crates/flui-tree/examples/basic_traversal.rs
@@ -34,7 +34,7 @@ impl DemoTree {
     }
 
     fn insert(&mut self, name: &str, parent: Option<ElementId>) -> ElementId {
-        let id = ElementId::new(self.nodes.len() as u64 + 1);
+        let id = ElementId::new(self.nodes.len() + 1);
         self.nodes.push(Some(DemoNode {
             name: name.to_string(),
             parent,
@@ -43,7 +43,7 @@ impl DemoTree {
 
         // Add to parent's children
         if let Some(parent_id) = parent {
-            if let Some(Some(parent_node)) = self.nodes.get_mut(parent_id.get() as usize - 1) {
+            if let Some(Some(parent_node)) = self.nodes.get_mut(parent_id.get() - 1) {
                 parent_node.children.push(id);
             }
         }
